\name{get_links}
\alias{get_links}
\title{wiki links from xml}
\usage{
  get_links(xml_files)
}
\arguments{
  \item{xml_files}{A vector containing plain character xml
  documents}
}
\value{
  A character vector with the links
}
\description{
  Extract Links from a Wikipedia file
}
\details{
  This functions extracts the links of an wikipedia
  article.
}
\author{
  chris
}

